# backend/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hotel Hall Booking Management System API",
    "version": "1.0.0",
    "description": "Documented per docs/03-standards/api-standards.md §16. One `paths`/`components.schemas` section grows per module as each is implemented — this file currently covers Authentication & Account Management (Module 1), Hotel Management (Module 3), and Hall Management (Module 4, partial — only the endpoints actually implemented so far)."
  },
  "servers": [{ "url": "/api/v1", "description": "Version 1 (api-standards.md §3)" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token, per api-standards.md §12: `Authorization: Bearer <JWT>`."
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "description": "api-standards.md §7 — every successful response.",
        "properties": {
          "status": { "type": "string", "enum": ["success"] },
          "message": { "type": "string" },
          "data": {},
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["status", "message", "data", "timestamp"]
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "api-standards.md §8 — every non-2xx response.",
        "properties": {
          "status": { "type": "string", "enum": ["error"] },
          "error": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "field": { "type": "string" }, "message": { "type": "string" } }
            }
          },
          "timestamp": { "type": "string", "format": "date-time" },
          "requestId": { "type": "string", "format": "uuid" }
        },
        "required": ["status", "error", "message", "timestamp", "requestId"]
      },
...
      "PublicHall": {
        "type": "object",
        "description": "The Hall entity (Technical Design §4). profileData's required/optional/custom structure is defined by BDR-016 (name, capacity required; description, location optional; any other key is a custom field) — the object itself remains unstructured JSON, no per-field schema properties. No visibility/status field is returned — visibility (Hidden/Visible) is computed live, never persisted (Technical Design §6), and its only observable effect is whether a request for a given Hall succeeds at all.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "hotelId": { "type": "string", "format": "uuid" },
          "profileData": { "type": "object", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "bookingTerms": { "type": "object", "properties": { "currency": { "type": "string", "enum": ["USD"] }, "rentAmountCents": { "type": "integer", "nullable": true }, "rentDurationHours": { "type": "integer", "enum": [24], "nullable": true }, "advancePaymentPercent": { "type": "number", "enum": [30], "nullable": true }, "customerServiceNumber": { "type": "string", "nullable": true }, "paymentReceivingNumber": { "type": "string", "nullable": true } } }
        }
      },
...
    }
  },
  "paths": {
...
    "/hotels/{hotelId}/halls": {
      "post": {
        "tags": ["Halls"],
        "summary": "Create a Hall belonging to a Hotel",
        "description": "HL1, HL2, BR-HALL-02 — Technical Design §11. Own-Hotel only. No precondition on the owning Hotel's own status — a Hall may be created whether the Hotel is REGISTERED, APPROVED_ACTIVE, or anything between. Per BDR-016 (Approved), profileData.name and profileData.capacity are required.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profileData": {
                    "type": "object",
                    "description": "BDR-016: name and capacity required; description, location, and any custom key are optional.",
                    "properties": {
                      "name": { "type": "string", "description": "Required — Hall Name." },
                      "capacity": { "type": "integer", "minimum": 1, "description": "Required — a valid positive integer." },
                      "description": { "type": "string", "description": "Optional." },
                      "location": { "type": "string", "description": "Optional — Location / Area." }
                    },
                    "required": ["name", "capacity"]
                  }
                },
                "required": ["profileData"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hall created.",
            "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "properties": { "data": { "$ref": "#/components/schemas/PublicHall" } } }] } } }
          },
          "400": { "description": "Malformed request shape, missing/empty Hall Name, or missing/invalid Capacity (BDR-016).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "401": { "description": "Missing, invalid, or expired access token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "404": { "description": "hotelId doesn't exist, or isn't owned by the caller (api-standards.md §13, tenant isolation — never 403).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      },
      "get": {
        "tags": ["Halls"],
        "summary": "List one Hotel's Halls",
        "description": "Technical Design §11. Public, with conditional behavior: the owning Hotel Manager sees every Hall regardless of visibility; anyone else sees the full page only while the Hotel is eligible (BR-HOTEL-04/05), or an empty page otherwise — every Hall under one Hotel shares the identical eligibility outcome (Technical Design §6). Offset pagination (coding-standards.md §6 — a bounded, per-Hotel list), distinct from the platform-wide GET /halls' cursor pagination.",
        "security": [{ "bearerAuth": [] }, {}],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "A page of Halls — every Hall if the caller is the owning Hotel Manager, otherwise only Visible ones (possibly an empty page).",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessEnvelope" },
                    {
                      "properties": {
                        "data": { "type": "array", "items": { "$ref": "#/components/schemas/PublicHall" } },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": { "type": "integer" },
                            "limit": { "type": "integer" },
                            "total": { "type": "integer" },
                            "totalPages": { "type": "integer" },
                            "hasNext": { "type": "boolean" },
                            "hasPrevious": { "type": "boolean" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "description": "Request validation failed (page/limit).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "404": { "description": "hotelId doesn't exist.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
    "/hotels/{hotelId}/halls/{id}": {
      "get": {
        "tags": ["Halls"],
        "summary": "Retrieve a single Hall",
        "description": "HL2, HL5, HL6 — Technical Design §11, §14.3. Public, with conditional behavior: if a valid token for the owning Hotel Manager is presented, the Hall is returned regardless of visibility; otherwise (no token, or a token for a different Hotel) it is returned only if Visible.",
        "security": [{ "bearerAuth": [] }, {}],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Hall retrieved.",
            "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "properties": { "data": { "$ref": "#/components/schemas/PublicHall" } } }] } } }
          },
          "404": { "description": "Hall doesn't exist, belongs to a different Hotel, or is Hidden and the caller isn't its owning Hotel Manager — never distinguished (api-standards.md §9, \"never leak existence\", generalized to visibility per Technical Design §6/§12).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      },
      "patch": {
        "tags": ["Halls"],
        "summary": "Update a Hall's profile information",
        "description": "HL3, BR-HALL-07 — Technical Design §11. Own-Hotel only. Every well-formed, authorized change applies immediately — no review step exists for Halls. If name or capacity is included, per BDR-016 it is content-validated (non-empty name, valid positive-integer capacity); every other key (description, location, custom fields) is merged unvalidated beyond shape.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "One or more profile fields to change; at least one field required.",
                "properties": {
                  "name": { "type": "string", "description": "Optional on update; if present, must be non-empty (BDR-016)." },
                  "capacity": { "type": "integer", "minimum": 1, "description": "Optional on update; if present, must be a valid positive integer (BDR-016)." },
                  "description": { "type": "string" },
                  "location": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hall updated.",
            "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "properties": { "data": { "$ref": "#/components/schemas/PublicHall" } } }] } } }
          },
          "400": { "description": "Malformed fields, an empty update body, or an included name/capacity that fails content validation (BDR-016).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "401": { "description": "Missing, invalid, or expired access token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "404": { "description": "Hall not found, or belongs to a Hotel the caller doesn't own (never 403).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
...
    "/halls": {
      "get": {
        "tags": ["Halls"],
        "summary": "Browse Visible Halls platform-wide",
        "description": "BR-HALL-08, BDR-009 — Technical Design §11 (added during that design, mirroring GET /hotels' own precedent). Public, unconditionally; never returns a Hidden Hall to anyone, including an authenticated Hotel Manager. Cursor pagination (coding-standards.md §6, Technical Design v1.2).",
        "parameters": [
          { "name": "hotelId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "A page of Visible Halls.",
            "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/PublicHall" } } } }] } } }
          },
          "400": { "description": "Request validation failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    }
  }
}